Remove workaround for #896019
The version of GLib in Debian 11 cleaned up a potential broken state
left behind by Debian 9, and Debian doesn't support skipping a major
release, so Debian 12 no longer needs this.

Similarly, the version in Ubuntu 22.04 LTS had this cleanup, and Ubuntu
doesn't support skipping a LTS release, so 22.10 and 24.04 LTS don't
need this.