d/rules: Increate arbitrary test timeout on all architectures except x86
If a test has a deadlock, we want it to fail reasonably quickly on the
architecture where developers are most likely to be building
interactively, so that test iteration is reasonably fast. However, if
a test is merely slow (perhaps while running under qemu) but would have
passed eventually, we still want it to pass eventually.

Previously we had lists of known-slow architectures, but it scales
better if we have a list of known-fast architectures, and assume that
everything else is (at least potentially) slow.

Thanks: Adrian Bunk