d/control: Take over gir1.2-glib-2.0{,-dev} from src:gobject-introspection
The GIR XML and typelibs were historically generated by the
gobject-introspection source package to avoid a circular dependency
between glib2.0 and gobject-introspection, but they were always
conceptually part of glib2.0 (and separating them like this has required
us to bundle a second entire copy of the glib2.0 source code in the
gobject-introspection source package, to address concerns about whether
the subset of that source code included in the upstream source package
can be said to be a preferred form for modification).

Now that build-profiles are well-understood, we can bootstrap new
architectures by initially building glib2.0 with the nogir build-profile,
then using that to build gobject-introspection, and then rebuilding
glib2.0 without nogir.

Alternatively, now that GObject-Introspection data can be cross-compiled,
we can bootstrap new architectures by cross-compiling glib2.0 and
gobject-introspection on an existing architecture (which must be of
the same endianness as a result of internal details).