d/control: Build-depend on libglib2.0-dev if introspection will be enabled
This works around a missing dependency in the gobject-introspection
package, for which a fix is pending. In practice other packages that
use gobject-introspection are highly likely to all depend on GLib
anyway, so the missing dependency likely only affects this one source
package.

We need this circular dependency for Meson to be able to discover
the gobject-introspection-1.0 pkg-config module successfully. As with
previous uses of GObject-Introspection, it can be bootstrapped by
cross-compiling or by using the nogir build profile (together with
nodoc, if building Architecture: all).