Build-depend on dpkg-dev (>= 1.22.5)
This is a safety catch against backports to suites that had the 32-bit
time_t ABI. If this version gets backported, then the whole t64 package
rename (including this change) will need to be reverted in the backport.

Not uploading this right now because it would delay the t64 transition,
but the change should be in the next upload.