d/libglib2.0-0t64.preinst: Remove libglib2.0-0 postrm to avoid file loss
During the migration from libglib2.0-0 to libglib2.0-0t64, the package
that is responsible for "owning" /usr/lib/*/gio/modules/giomodule.cache
and /usr/share/glib-2.0/schemas/gschemas.compiled changed from
libglib2.0-0 to libglib2.0-0t64. Because dpkg does not have an equivalent
of RPM's %ghost files, the ownership of these files is managed by social
convention rather than by the package management system.

Unfortunately, libglib2.0-0's postrm as shipped in Debian releases from
2010 to the present is not aware of the possibility that another binary
package might need to take over responsibility for those files, and so
will remove both files during purge (and giomodules.cache also during
remove) in accordance with the requirement that installing and then
removing and purging a package must not leave unowned files behind.
This causes most applications that depend on GSettings schemas to crash
with an assertion failure, until the next time the glib-compile-schemas
trigger happens to be run; it will also cause functionality loss for
applications that depend on GIO modules.

To disarm the problematic maintainer script, delete it during the new
package's preinst. This is (probably) a Policy violation, but seems like
the least-bad exit strategy from the unacceptable situation we have
found ourselves in.

A subsequent commit will improve the postrm so that if we find that we
need to migrate from libglib2.0-0t64 to libglib2.0-0xyz or libglib-2.0-0
at some point in the future, similar efforts will not be needed.

Closes: #1065022