d/tests: Add a manual reproducer for #1065022
Sorry, I do not have sufficient spoons right now to make this into a
proper autopkgtest so that it can be run for every upload. My experience
from doing similar things in nss-mdns is that an autopkgtest that goes
back in time and installs previous suites' packages comes with a
significant maintenance cost.