d/tests/1065022-futureproofing: Add a test for recurrence of #1065022
If at some point in the future we have another transition as extensive
as time64, then libglib2.0-0t64 could conceivably be replaced by some
other package, which I have modelled here as libglib2.0-0xyz. If that
happens, we need to avoid deletion of gschemas.compiled and
giomodule.cache, otherwise we will have another bug similar to #1065022.

This test-case depends on several implementation details of
dpkg-repack and libglib2.0-0t64, and will likely fail in a non-minimal
chroot/container (if there are foreign-architecture instances of
libglib2.0-0t64), so it might need to be adjusted in the future. As
a result, I have marked it as flaky, so that failures in the official
autopkgtest environment will not be considered a release-critical bug
that stalls migration and requires immediate intervention by maintainers.

(cherry picked from commit 6271980cc8f9366b956de0d26b0ba357bc858e77)