d/libglib2.0-0t64.postrm: Only clean up giomodule.cache during purge
This matches the behaviour that we have had for gschemas.compiled since
commit c8b28bc3 "debian/libglib2.0-0.postrm.in: Only remove the compiled
schemas on purge", and with similar reasoning: if we remove this file
during remove, then during upgrades there is a window between old-postrm
and new-postinst during which giomodule.cache is missing. This will result
in applications that require a GIO module (for example for TLS, proxy
discovery or non-local filesystems) failing to load it, which seems
undesirable.

The impact is less serious than for GSettings schemas, but if we can
avoid it, it seems like we should.