d/libglib2.0-0t64.postrm: Avoid recurrence of #1065022 in the future
If at some point in the future we have another transition as extensive
as time64, then libglib2.0-0t64 could conceivably be replaced by some
other package, which I have modelled here as libglib2.0-0xyz. If that
happens, we need to avoid deletion of gschemas.compiled and
giomodule.cache, otherwise we will have another bug similar to #1065022.

This implementation is based on the assumption that GLib will most
likely continue to ship helper executables in ${libdir}/glib-2.0; or that
if upstream carries out a major ABI break (unlikely to happen soon due
to its extensive impact), they will make it fully parallel-installable,
with ${datadir}/glib-3.0/schemas and ${libdir}/gio-3.0/modules, in
which case the old version can continue to maintain or delete its caches
without impacting the new version.