d/control: libglib2.0-dev Suggests gir1.2-glib-2.0-dev
Previously libgirepository1.0-dev was suggested, but that package is not
Multi-Arch: same. The reason for the Suggests was that it used to be the
only way to get the GIR XML for GLib-2.0 and related modules, which is
often part of the -dev package. Now that we have gir1.2-glib-2.0-dev
available, it's a better choice for this purpose.