Use host-architecture gi-compile-repository, etc. when cross-compiling
When setting up the cross wrappers for gi-compile-repository, etc. I had
assumed that gi-compile-repository was a simple transformation from
GIR XML into binary, which varied only by its endianness. Unfortunately,
it is not: it also transforms abstract types such as size_t into
equivalent fixed-size types such as guint64, which requires knowledge
of the size of each type.

Instead of running the build architecture gi-compile-repository and
telling it to use the host architecture search path, install upstream's
gi-compile-repository etc. into ${pkglibdir}, and set up cross wrappers
that will automatically detect whether we can run them directly or
whether we must use qemu-user. This is not "pure" cross-compilation,
but it's the best we can do.

This means that the build and host OS must match, but it does have the
side benefit that we no longer require the build and host endianness to
be the same, because everything that interacts the typelib is now a
host binary (possibly running under qemu).

Closes: #1066904