d/control: Bring back libglib2.0-0 as a transitional package
On architectures that were unaffected by the 64-bit time_t transition,
there was no ABI break, so we can upgrade smoothly from the old to new
name. This might help to hint apt into doing the upgrade, in preference
to removing packages we wanted to keep.

On architectures that were affected by the transition (mainly armel
and armhf), it would be incorrect for this package to exist, so it is
not built.