Move libglib2.0-dev-bin, except gi-*, to libgio-2.0-dev-bin
The versions of libglib2.0-dev and -dev-bin in Debian testing/unstable
and Ubuntu 24.04 depend on python3 from the host architecture or
qemu-user from the build architecture (for the wrapped gi-* tools),
which is problematic for bootstrapping. Separate out the rest, so that
packages that participate in bootstrapping can be compiled with only
libgio-2.0-dev-bin (possibly using the nogir build-profile).

Autoconf macros and gettext rules move to libgio-2.0-dev instead, since
they are not executables.

Helps: #1078929