Move ${multiarch}-gi-* tools to girepository-tools
This converts libglib2.0-dev into a pure metapackage, with no functionally
significant content of its own, except for the gi-* symlinks to the
primary architecture's GObject-Introspection-related tools.

libglib2.0-dev continues to be appropriate to use as a build-dependency
in most cases. For low-level packages that are involved in architecture
bootstrapping, consider using libgio-2.0-dev instead.

Helps: #1078929