d/control: libgio-2.0-dev-bin Breaks/Replaces old libglib2.0-dev
/usr/share/glib-2.0/gdb/*_gdb.py were part of libglib2.0-dev in
bookworm, then moved to libgio-2.0-dev as part of fixing #1078929,
then moved from there to libgio-2.0-dev-bin to fix #987802. We already
had all of the Breaks/Replaces necessary to do each of those smaller
upgrades successfully, but because Replaces is not a transitive
relationship, we need a third set of Replaces for reliable upgrades
from bookworm to trixie.

Closes: #1100142
Thanks: Helmut Grohne