d/patches: Ensure that tzset() is called before localtime_r()
Otherwise the behaviour of localtime_r() is unspecified.

This brings us up to date with upstream commit 2.84.1-15-gb3de15acf9,
excluding changes that are not relevant to Debian architectures
(macOS CI and Windows) and the revert of the fix for LP#2072586
(which was reverted in the upstream 2.84.x branch as a behaviour change,
but kept in upstream 2.85.x).