Generate a dependency on a virtual package for libffi-related symbols
Every time the ABI of libffi breaks, any dependent package that uses
libgirepository's libffi-dependent functionality (in practice this means
dynamic language bindings like PyGI and gjs) will need to be rebuilt
against a version of libgirepository that has, itself, been rebuilt
against the new libffi. To keep track of this, generate a virtual package
like libgirepository-2.0-0-with-libffi8 and make libffi-ABI-dependent
symbols like gi_function_info_prep_invoker() generate a dependency on it.

This matches how we handled #995032 in libgirepository-1.0-1.

Closes: #1110825