postrm: Don't remove cache files if they would be non-empty
This guards against issues similar to #1065022 and #1110696. If we
delete giomodule.cache while there is still at least one GIO module,
that will break GIO module loading, and similarly if we delete
gschemas.compiled while there is at least one GSettings schema, that
will break GSettings schema loading (which is particularly
high-visibility because it results in applications crashing with a
fatal error).