preinst: Disarm libglib2.0-0 postrm for all architectures
Previously we used dpkg --print-architecture and
dpkg --print-foreign-architectures to enumerate all of the architectures
for which dpkg would allow installing new packages, but that is not
actually enough: it is possible for a foreign architecture to be
disabled, but still have packages from that architecture installed in
a removed-but-not-purged state. In that scenario, the old postrm would
still be harmful, causing collateral damage to other architectures by
deleting gschemas.compiled when the package is subsequently purged.

Instead we ask dpkg to list all of the architectures of libglib2.0-0,
which includes architectures for which it was removed-but-not-purged,
and remove their faulty postrm scripts.

Closes: #1110696