debian/control: Add a Breaks to gir1.2-glib-2.0 on gnome-shell << 49~rc
GNOME Shell 49~rc was the first version that properly used GioUnix
instead of Gio namespace for DesktopAppIcon and that builds its typelib
with it as dependency.

So break on that. We use the ubuntu version as it should not affect
debian at all, but avoids a light fork just for this.